+++ src/main/kotlin/com/github/malbanese93/Main.kt
@@ -1,8 +1,9 @@
 package com.github.malbanese93
 
+import com.github.malbanese93.chip8.Chip8
+
 fun main() {
     val chip8 = Chip8()
-
     chip8.start()
 }
